Move http code to package
