+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/riraum/si-cheong/v2
+module github.com/riraum/si-cheong
 
 go 1.24.1
